flask_app/requirements.txt
flask
flask-cors
kafka-python
gunicorn